backend/requirements.txt
# Web Framework
fastapi>=0.110.0,<1.0.0
uvicorn[standard]>=0.27.0,<1.0.0

# Database
sqlalchemy>=2.0.25,<3.0.0
greenlet>=3.0.0,<4.0.0
asyncpg>=0.29.0,<1.0.0

# AI
anthropic>=0.40.0,<1.0.0

# PDF Generation
reportlab>=4.1.0,<5.0.0
Pillow>=10.0.0,<11.0.0

# Config & Validation
python-dotenv>=1.0.0,<2.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
python-multipart>=0.0.7,<1.0.0

# Auth
python-jose[cryptography]>=3.3.0,<4.0.0
passlib[bcrypt]>=1.7.0,<2.0.0

# Testing
pytest>=8.0.0,<9.0.0
pytest-asyncio>=0.23.0,<1.0.0
httpx>=0.27.0,<1.0.0
